requirements/pipeline.requirements.txt
polars==1.22.0
google-generativeai==0.8.4
PyYAML==6.0.2
pymongo==4.11.1
grpcio==1.67.1